Final Exam Code
Finally here is the code of the problem 3A and 3B of the final exam! For
the rest the code was already provided and they were purely questions to
test your understanding so I didn't uploaded them. I hope you had fun
throughout this course and that I helped you learn :-)